Remove RAlt as a modifier in non-iso layouts.
In ABNT2 brazilian portuguese keyboard layout RAlt (or AltGr in ABNT2)
does not work the same as LAlt and should not be used as a modifier in
Neovim. It is used to enable typing special characters in combination
with Shift, such as ?`´~^ºª etc, either as the only way for typing them
or as a faster way combining with other keys.
Using Neovim in xterm, kitty, gnome-terminal etc has the expected
behavior, as using in other places in the system as well (e.g. browsers)