Put Windows artefact in zip archive
This PR adds an extra step to the Windows pipeline to put the Windows artefact in a `.zip` archive. 

A lot of places (specially companies) has proxies that outright blocks everything ending in `.exe` and this is a neat way to make it a bit easier to distribute from Github to those places.