updates rmpv (0.4.4 -> 1.0.0)
When trying to build neovide via `cargo install --git
https://github.com/neovide/neovide` it failed.
The reason was, that the `--git` option by default ignores the
Cargo.lock file and resolves crates version on-demand.

Probably because the master branch of `nvim-rs` tracked, there was
a version mismatch of rmpv that it and neovide used.

Bumped version of rmpv in neovide to 1.0.0 to match nvim-rs.
Refreshed the Cargo.lock file via `cargo update`.
Also, fixed a warning where a variable need not be mut.