Added alternative Alt + character handing for MacOS
By default on MacOS pressing a key combination containing an Alt key
produces a special character:

    <Alt + s> -> ß
    <Alt + v> -> √
    ...

It is really unfortunate if you want to use those key combinations as
shortcuts and this case has to be handles separately by the app.

This commit introduces a new configuration flag
`neovide_macos_alt_is_meta`. When the flag is set to true then all the
Alt- key combinations are treated as shortcuts with no extra character
modifications, i.e. pressing `Alt` and `s` indeed emits a <M-s>
shortcut.