Fix #1207
The original check of whether a cursor movement was intentional or just
in the same window wasn't enough for the command line, where a cursor
didn't really move but was just displayed in the command line area for a
short time. Hardcodes 4 as constant for the commandline mode, I have no
idea how future-safe that is.