Make x11 WM_CLASS instance configurable
It was considered to merge both parts of the property (instance and
class) into just one configuration option. This would better
reflected the inherent nature of WM_CLASS, which is also just one
option and not two.
Still, this approach was not taken because it would break backwards
compatibility, and given the low benefit it is not worth doing so.