fix: Change how window resize detection works (#1458)
Changes how the detection of window resizing works. Previously relied
on hardcoded 800 x 600 value, which was not always correct on MacOS, or
rescaling by the scaling factor, which was not always correct on Linux.

The physical size of the window is now stored right after the window is
initialized, and this value is used whenever we check for window
resizing.