Handle clipboard stores via event aggregator.
The X11 clipboard doesn't work correctly if we set the clipboard
directly from the handler responsible for the neovide.set_clipboard RPC.
I'm not certain why this is, but I presume it has to do with the thread
handling the request being rather short-lived. To solve the problem, we
delegate this task to the clipboard command handler, which is run by a
long-running thread.