add quake mode accessibility like other terminals
adds a new section to the faq page of the project explaining how one can achieve quake mode accessibility like other terminals using hammerspoon.

includes code snippets for adding a key binding to quickly open neovide on the current space. this is done by adding the code to `~/.hammerspoon/init.lua` and binding a specific hotkey to the function. this function will then check if the app is already open and if so, will either open a new window, hide it, or move it to the current space and bring it to the front. finally, if the app is not open, it will open it.