Make window title customizable
Let the user customize the default “Neovide” window title via
“--window-title” command-line arguments option.

This change is addressing https://github.com/neovide/neovide/issues/1904
but it doesn’t fully resolve it. It’s still to figure out how to set
WM_NAME to make the feature complete for X11. Most of the X11
applications seems to set both the title and the WM_NAME property.