[macOS input] fix option-as-meta
Here we re-implement support for option-as-meta by using the winit API:
    window.set_option_as_alt(oa)
...where oa is of type winit::platform::macos::OptionAsAlt.

By relying on the winit API this way, we are now able to map only the left, only
the right, or both of the two physical 'option' keys to the virtual key 'meta',
which leaves the other 'option' key free so that it can be used to access the
additional two layers of the input source. (E.g. with the u.s. english input
source, option-c produces "ç" and option-shift-c produces "Ç".) This is done by
setting the nvim variable:

        vim.g.neovide_input_macos_option_key_is_meta

Possible values (defined by init::platform::macos::OptionAsAlt) are:
    "Both"
    "OnlyLeft"
    "OnlyRight"
    "None"

The preëxisting variable vim.g.neovide_input_macos_alt_is_meta can still be
used, and if true, it will internally result in the use of OptionAsAlt::Both.

Separately, this commit fixes an issue where, in the event of a key-press like
option-shift-command-y (with vim.g.neovide_input_macos_alt_is_meta set to true),
neovide would produce the string "<M-S-D-y>" instead of "<M-D-Y>" because the
`option` key was down, and format_normal_key() took this to imply that the key
is "special" (i.e. in the same category as escape, return, F1, etc. as opposed
to the category of printable characters), which led format_modifier_string() to
include "S-" in its output.

Also, in format_modifier_string(), the test for whether 'meta' is pressed has
been simplified.