fix: prevent empty window title
If Neovim attempts to send an empty window title, it is now set to
Neovide instead, rather than the empty string.

This improves two user scenarios:

- Users who utilize assistive technologies will now always observe a
  window title by default.
- Users of window managers that rely on window titles will have improved
  integration.

The precipitating use case was using Komorebi, a tiling window manager
for Windows. Komorebi uses an empty window title as a strong signal that
the software is a hidden window, as this is a common convention on the
platform.

Updates #1553