Don't set the IME area until the IME system accepts it
Also throttle the requests.

Before these changes fcitx5 would crash on Wayland all the time, so
setting it every frame was clearly not the correct way to fix the
problem. I guess the real issue was that setting the area when it was
not allowed caused the IME system to go into an invalid state.