add --chdir option
Start neovim in the given working directory. This is particularly useful
in a situation where the caller does not launch neovide directly itself,
but does provide the arguments to be used for launching. For example,
consider this macOS `open` command invocation:

  open -b com.neovide.neovide --args --chdir ${PWD} foo.txt

Note that macOS launches applications with `/` as their current working
directory. Without specifying `--chdir`, neovim would attempt to open
a non-existent file `/foo.txt`.