fix: native rendering for box drawing chars (#3033)
* fix: native rendering for box drawing chars

This PR adds optional native rendering of box drawing chars so that
there are no gaps between edges, breaking continuity. The current set of
characters is not complete, but this lays the foundation to add more
codepoints and their renderings as needed.

fixes: #2491, #2854

rebase fixes

* review comments + lint fixes

* fixed docs lint errors + test file for visually checking alignment

* Remove backgroud drawing

* Draw box characters on a separate layer

* Fixup the antialias settings

* Round the antialiased line start and endpoints

* Fix double line drawing

* Fix dashed lines

* Restore the blend mode

* Refactor double line drawing

* Fix double line joints

* Fix the color of vertically dashed line

* Fix the full diagonal rendering

* Render the box glyphs to absolute coordinates

This ensures that the pixel rounding is correct both for antialiased and
non-antialiased lines. Note there can be a slight transformation when
animating.

* Fix rounded corner drawing

* Draw perfectly aligned semicircles

* Remove ◖and ◗

Both can be rendered by the normal font, since they don't extend outside
of the box. They were also rendered wrong before, they should not take
up the full width, since they can be paired with glyps like ◕

* Fix line alignment

* Fix arrow alignment

* More alignment fixes

* Allow more arguments to draw_foreground

* Add more test examples

* Remove unneeded color parameter

* Add missing corner types

This also makes the draw_line function more generic

* Add missing cross junctions

* Add missing glyphs to the test, and better formatting

* Add missing dashed line

* Remove unicode arrows

* Add some powerline tests

* Nicer dashed line rendering

* Fix minor alignment error

* Fix shade drawing

* Improve the diagonal fill quality by drawing slightly outside the clipbox

* Remove extra layer

The boxchars are drawn fine without it

* Fix clippy

* Remove points_to_pixels logspam

* Use the same fractional coordinates as the target position for animated or moved floating windows

Otherwise, there will be glitches and gaps, since the box drawing
characters are drawn using the target positions.

* added e0d2 and e0d4

* added per-size line thickness

* adding stuff back

* updated docs

* Remove logspam

* Update the defaults

* Update the documentation

* Fix markdown linting

* Fix link

---------

Co-authored-by: Fred Sundvik <fsundvik@gmail.com>